requirements.txt
azure-functions
fastapi
azure-identity
azure-storage-blob